+++ src/settingsList.lua
@@ -3,10 +3,12 @@ return {
     id = 'CheckUpdates',
     defaultValue = true,
     label = 'Check for updates',
+    description = 'If checked, the plugin will check for updates at Studio launch and display a notice on the plugin when outdated. If unchecked, the plugin will not for any updates.'
   },
   {
     id = 'DeletePart',
     defaultValue = true,
     label = 'Delete part when converted',
+    description = 'If checked, the part used to be converted will be delete. If unchecked, it will stay in the Workspace when converted.'
   }
 }